dashboard/requirements.txt
streamlit==1.31.0
psycopg2-binary==2.9.9
pandas==2.2.0
plotly==5.18.0